:recycle: VectorMemoryRepositoryの実装をインフラレイヤーに移動
