+++ src/infrastructures/external_services.rs
@@ -1,1 +1,2 @@
 pub(crate) mod openai_embedding_repository;
+pub(crate) mod qdrant_vector_memory_repository;
